+++ lithops/constants.py
@@ -106,6 +106,7 @@ SERVERLESS_BACKENDS = [
     'azure_functions',
     'azure_containers',
     'aliyun_fc',
+    'oracle_f',
     'k8s'
 ]
 
@@ -126,7 +127,8 @@ FAAS_BACKENDS = [
     'cloudrun',
     'azure_functions',
     'azure_containers',
-    'aliyun_fc'
+    'aliyun_fc',
+    'oracle_f'
 ]
 
 BATCH_BACKENDS = [