+++ lithops/invokers.py
@@ -28,8 +28,7 @@ from lithops.future import ResponseFuture
 from lithops.config import extract_storage_config
 from lithops.version import __version__
 from lithops.utils import verify_runtime_name, version_str, is_lithops_worker, iterchunks
-from lithops.constants import LOGGER_LEVEL, LOGS_DIR,\
-    LOCALHOST, SERVERLESS, STANDALONE
+from lithops.constants import LOGGER_LEVEL, LOGS_DIR, SERVERLESS
 from lithops.util.metrics import PrometheusExporter
 
 logger = logging.getLogger(__name__)
@@ -406,6 +405,8 @@ class FaaSInvoker(Invoker):
         Normal Invocation
         Use local threads to perform all the function invocations
         """
+        self.compute_handler.pre_invoke(job)
+
         if self.remote_invoker:
             return self._invoke_job_remote(job)