+++ lithops/serverless/backends/aliyun_fc/aliyun_fc.py
@@ -37,7 +37,7 @@ class AliyunFunctionComputeBackend:
     A wrap-up around Aliyun Function Compute backend.
     """
 
-    def __init__(self, afc_config, storage_config):
+    def __init__(self, afc_config, internal_storage):
         logger.debug("Creating Aliyun Function Compute client")
         self.name = 'aliyun_fc'
         self.type = 'faas'
@@ -226,7 +226,7 @@ class AliyunFunctionComputeBackend:
     def list_runtimes(self, runtime_name='all'):
         """
         List all the runtimes deployed in the Aliyun FC service
-        return: list of tuples (docker_image_name, memory)
+        return: list of tuples (container_image_name, memory, version)
         """
         logger.debug('Listing deployed runtimes')
         runtimes = []
@@ -244,7 +244,7 @@ class AliyunFunctionComputeBackend:
                     runtimes.append((name, memory, version))
         return runtimes
 
-    def invoke(self, runtime_name, memory=None, payload={}):
+    def invoke(self, runtime_name, memory, payload={}):
         """
         Invoke function
         """