+++ lithops/serverless/backends/oracle_f/entry_point.py
@@ -1,5 +1,4 @@
-#
-# Copyright Cloudlab URV 2021
+# (C) Copyright Cloudlab URV 2023
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,32 +14,39 @@
 #
 
 import os
+import io
 import json
 import logging
+
 from lithops.version import __version__
 from lithops.utils import setup_lithops_logger
 from lithops.worker import function_handler
 from lithops.worker import function_invoker
 from lithops.worker.utils import get_runtime_metadata
 
-logger = logging.getLogger('lithops.worker')
+
+logger = logging.getLogger(__name__)
 
 
-def main(event, context):
-    args = json.loads(event)
-    os.environ['__LITHOPS_ACTIVATION_ID'] = context.request_id
-    os.environ['__LITHOPS_BACKEND'] = 'Aliyun Function Compute'
+def handler(ctx, data: io.BytesIO = None):
+
+    args = json.loads(data.getvalue())
+    call_id = ctx.CallID()
+
+    os.environ['__LITHOPS_ACTIVATION_ID'] = call_id
+    os.environ['__LITHOPS_BACKEND'] = 'Oracle Functions'
 
     setup_lithops_logger(args.get('log_level', logging.INFO))
 
     if 'get_metadata' in args:
         logger.info(f"Lithops v{__version__} - Generating metadata")
-        return get_runtime_metadata()
+        metadata = get_runtime_metadata()
+        return metadata
     elif 'remote_invoker' in args:
-        logger.info(f"Lithops v{__version__} - Starting Aliyun Function Compute invoker")
+        logger.info(f"Lithops v{__version__} - Starting Oracle Functions invoker")
         function_invoker(args)
     else:
-        logger.info(f"Lithops v{__version__} - Starting Aliyun Function Compute execution")
+        logger.info(f"Lithops v{__version__} - Starting Oracle Functions execution")
         function_handler(args)
 
-    return {"Execution": "Finished"}
+    return call_id