+++ lithops/storage/backends/aws_s3/aws_s3.py
@@ -102,7 +102,7 @@ class S3Backend:
 
     def create_bucket(self, bucket_name):
         """
-        Create a bucket if not exists
+        Create a bucket if it doesn't exist
         """
         try:
             self.s3_client.head_bucket(Bucket=bucket_name)