+++ lithops/storage/backends/gcp_storage/gcp_storage.py
@@ -53,7 +53,7 @@ class GCPStorageBackend:
 
     def create_bucket(self, bucket_name):
         """
-        Create a bucket if not exists
+        Create a bucket if it doesn't exist
         """
         try:
             bucket = self.client.bucket(bucket_name)